% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{load_file}
\alias{load_file}
\title{Load a GTF or GFF3 file from GENCODE as a data frame.}
\usage{
load_file(filename)
}
\arguments{
\item{filename}{A character string representing the path to the GTF or GFF3 file (e.g., "gencode.vM36.annotation.gtf.gz").
The file could be in GTF or GFF3 format and must be downloaded from a reliable source like GENCODE.}
}
\value{
A data frame containing the parsed content of the GTF or GFF3 file.
The data frame includes standard columns such as 'seqnames', 'start', 'end', 'strand', 'feature', and 'gene_id', among others.
}
\description{
This function imports a GTF or GFF3 file (commonly from the GENCODE website) and converts it into a data frame.
The function provides flexibility for users to work with genomic feature files easily in the R environment.
}
\details{
The function uses the \code{rtracklayer} package to import the GTF or GFF3 file and returns it as a data frame.
The user should ensure that the input file is properly formatted and accessible from the specified path.
Files larger than a few hundred MBs may take longer to load and process.
}
\examples{
\dontrun{
# Load a GTF file from GENCODE
mouse_v36_gtf <- load_file(filename = "gencode.vM36.annotation.gtf.gz")
head(mouse_v36_gtf)  # Preview the first few rows of the data frame

# Load a GFF3 file if required
df_gff3 <- load_file("example.gff3")
head(df_gff3)  # Preview the first few rows of the data frame
}
}
\keyword{GENCODE,}
\keyword{GFF3}
\keyword{GTF,}
\keyword{import,}
